fix: install `sox` so FS sounds are correctly overridden
`sox` is used by the bbb-freeswitch-sounds package to reduce the volume
of Senfcall's mute/unmute sounds. Since it's installation was
deactivated due to build issues a while ago, the volume reduction step
was failing and sounds were not overwritten.

Install `sox` again so that FS sounds are correctly overridden.
Not sure why it was failing a while back, but it seems fine now. Tested
with a local image + a local build of bbb-freeswitch-sounds only.